+++ .nvmrc
@@ -1,1 +1,1 @@
-v18.19.0
+v18.19.1
