+++ package.json
@@ -46,14 +46,14 @@
         "lint"
     ],
     "dependencies": {
-        "@angular/animations": "~17.3.6",
-        "@angular/cdk": "~17.3.6",
-        "@angular/common": "~17.3.6",
-        "@angular/core": "~17.3.6",
-        "@angular/forms": "~17.3.6",
-        "@angular/platform-browser": "~17.3.6",
-        "@angular/platform-browser-dynamic": "~17.3.6",
-        "@angular/router": "~17.3.6",
+        "@angular/animations": "^19.0.0",
+        "@angular/cdk": "^19.0.0",
+        "@angular/common": "^19.0.0",
+        "@angular/core": "^19.0.0",
+        "@angular/forms": "^19.0.0",
+        "@angular/platform-browser": "^19.0.0",
+        "@angular/platform-browser-dynamic": "^19.0.0",
+        "@angular/router": "^19.0.0",
         "@capacitor-community/electron": "^5.0.1",
         "@capacitor/app": "5.0.6",
         "@capacitor/browser": "^5.1.0",
@@ -66,37 +66,39 @@
         "@ionic/storage-angular": "^4.0.0",
         "@ngneat/transloco": "^6.0.4",
         "@ngneat/until-destroy": "^10.0.0",
-        "@ngrx/effects": "17.2.0",
-        "@ngrx/store": "17.2.0",
+        "@ngrx/effects": "^19.0.0",
+        "@ngrx/operators": "^19.2.1",
+        "@ngrx/store": "^19.0.0",
         "date-fns": "^4.1.0",
         "deep-object-diff": "^1.1.9",
+        "ionicons": "^8.0.8",
         "ng-let": "^18.0.1",
-        "ngrx-store-localstorage": "^17.0.0",
+        "ngrx-store-localstorage": "19.0.0",
         "ngx-scrollbar": "^16.1.1",
         "ngx-tippy-wrapper": "^6.2.2",
-        "ngx-visibility": "~1.5.0",
+        "ngx-visibility": "1.8.0",
         "rxjs": "~7.8.1",
         "tslib": "^2.6.2",
-        "zone.js": "~0.14.3"
+        "zone.js": "~0.15.0"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~17.3.6",
-        "@angular-eslint/builder": "~17.3.0",
-        "@angular-eslint/eslint-plugin": "~17.3.0",
-        "@angular-eslint/eslint-plugin-template": "~17.3.0",
-        "@angular-eslint/schematics": "~17.3.0",
-        "@angular-eslint/template-parser": "~17.3.0",
-        "@angular/cli": "~17.3.6",
-        "@angular/compiler": "~17.3.6",
-        "@angular/compiler-cli": "~17.3.6",
-        "@angular/language-service": "~17.3.6",
+        "@angular-devkit/build-angular": "^19.0.0",
+        "@angular-eslint/builder": "^19.0.0",
+        "@angular-eslint/eslint-plugin": "^19.0.0",
+        "@angular-eslint/eslint-plugin-template": "^19.0.0",
+        "@angular-eslint/schematics": "^19.0.0",
+        "@angular-eslint/template-parser": "^19.0.0",
+        "@angular/cli": "^19.0.0",
+        "@angular/compiler": "^19.0.0",
+        "@angular/compiler-cli": "^19.0.0",
+        "@angular/language-service": "^19.0.0",
         "@capacitor/cli": "^5.6.0",
         "@digitak/esrun": "^3.2.26",
         "@ionic/angular-toolkit": "^11.0.1",
         "@ionic/cli": "^7.2.1",
-        "@ngrx/eslint-plugin": "17.2.0",
-        "@ngrx/schematics": "17.2.0",
-        "@ngrx/store-devtools": "17.2.0",
+        "@ngrx/eslint-plugin": "^19.0.0",
+        "@ngrx/schematics": "^19.0.0",
+        "@ngrx/store-devtools": "^19.0.0",
         "@types/chrome": "0.0.258",
         "@types/jasmine": "~5.1.4",
         "@types/jasminewd2": "~2.0.13",
@@ -126,7 +128,7 @@
         "protractor": "~7.0.0",
         "terser-webpack-plugin": "^5.3.10",
         "ts-node": "^10.9.2",
-        "typescript": "^5.3.3",
+        "typescript": "~5.8.3",
         "webpack": "^5.89.0",
         "webpack-bundle-analyzer": "^4.10.1"
     }
