+++ src/app/core/components/searchbar-results/searchbar-results.component.ts
@@ -1,3 +1,4 @@
+import { AsyncPipe, DatePipe, NgFor, NgIf, UpperCasePipe } from '@angular/common';
 import {
     ChangeDetectionStrategy,
     Component,
@@ -7,17 +8,45 @@ import {
     Output,
     ViewEncapsulation,
 } from '@angular/core';
+import {
+    IonItem,
+    IonLabel,
+    IonList,
+    IonNote,
+    IonText,
+} from '@ionic/angular/standalone';
+import { RouterLink } from '@angular/router';
+import { TranslocoPipe } from '@ngneat/transloco';
 
+import { ImageCardComponent } from '@app/shared/components/image-card/image-card.component';
 import { Observable } from 'rxjs';
 import { ResultOpenTarget, SearchbarResult } from '@app/shared/types/searchbar.types';
 import { SHIKIMORI_DOMAIN_TOKEN } from '@app/core/providers/shikimori-domain';
+import { ShikimoriMediaNamePipe } from '@app/shared/pipes/shikimori-media-name/shikimori-media-name.pipe';
 import { ShikimoriMediaNameType } from '@app/shared/types/shikimori/shikimori-media-name.type';
+import { SkeletonBlockComponent } from '@app/shared/components/skeleton-block/skeleton-block.component';
 import { trackById } from '@app/shared/utils/common-ngfor-tracking';
 
 @Component({
     selector: 'app-searchbar-results',
     templateUrl: './searchbar-results.component.html',
     styleUrls: ['./searchbar-results.component.scss'],
+    imports: [
+        IonList,
+        IonItem,
+        IonLabel,
+        IonNote,
+        IonText,
+        NgIf,
+        AsyncPipe,
+        DatePipe,
+        RouterLink,
+        UpperCasePipe,
+        TranslocoPipe,
+        ImageCardComponent,
+        SkeletonBlockComponent,
+        ShikimoriMediaNamePipe,
+    ],
     changeDetection: ChangeDetectionStrategy.OnPush,
     encapsulation: ViewEncapsulation.None,
 })
