+++ src/app/modules/contributions/contributions.routes.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+
+import { ContributionsPage } from '@app/modules/contributions/contributions.page';
+
+export const CONTRIBUTIONS_ROUTES: Routes = [
+    {
+        path: '',
+        component: ContributionsPage,
+    },
+];
