+++ src/app/modules/home/components/card-grid/card-grid.component.ts
@@ -1,3 +1,4 @@
+import { AsyncPipe, NgFor, NgIf, NgTemplateOutlet } from '@angular/common';
 import {
     ChangeDetectionStrategy,
     Component,
@@ -9,7 +10,10 @@ import {
 import { Observable } from 'rxjs';
 import { TranslocoService } from '@ngneat/transloco';
 
+import { CardGridItemComponent } from '@app/modules/home/components/card-grid-item/card-grid-item.component';
+import { GetPlayerLinkPipe } from '@app/shared/pipes/get-player-link/get-player-link.pipe';
 import { SHIKIMORI_DOMAIN_TOKEN } from '@app/core/providers/shikimori-domain';
+import { SkeletonBlockComponent } from '@app/shared/components/skeleton-block/skeleton-block.component';
 import { UserAnimeRate } from '@app/shared/types/shikimori/user-anime-rate';
 import { getAnimeName } from '@app/shared/utils/get-anime-name.function';
 import { trackById } from '@app/shared/utils/common-ngfor-tracking';
@@ -18,6 +22,15 @@ import { trackById } from '@app/shared/utils/common-ngfor-tracking';
     selector: 'app-card-grid',
     templateUrl: './card-grid.component.html',
     styleUrls: ['./card-grid.component.scss'],
+    imports: [
+        NgIf,
+        NgFor,
+        NgTemplateOutlet,
+        AsyncPipe,
+        SkeletonBlockComponent,
+        CardGridItemComponent,
+        GetPlayerLinkPipe,
+    ],
     changeDetection: ChangeDetectionStrategy.OnPush,
     encapsulation: ViewEncapsulation.None,
 })
