--- src/app/modules/home/home-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { HomePage } from '@app/modules/home/home.page';
-
-const routes: Routes = [
-    {
-        path: '',
-        component: HomePage,
-    },
-];
-
-@NgModule({
-    imports: [RouterModule.forChild(routes)],
-    exports: [RouterModule],
-})
-export class HomePageRoutingModule {}
