--- src/app/modules/settings/settings-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { SettingsPage } from '@app/modules/settings/settings.page';
-
-const routes: Routes = [
-    {
-        path: '',
-        component: SettingsPage,
-    },
-];
-
-@NgModule({
-    imports: [RouterModule.forChild(routes)],
-    exports: [RouterModule],
-})
-export class SettingsRoutingModule {}
