+++ src/app/modules/settings/settings.module.ts
@@ -1,14 +1,10 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { SettingsRoutingModule } from '@app/modules/settings/settings-routing.module';
-
-
 @NgModule({
     declarations: [],
     imports: [
         CommonModule,
-        SettingsRoutingModule,
     ],
 })
 export class SettingsModule {}
