+++ src/app/modules/settings/settings.routes.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+
+import { SettingsPage } from '@app/modules/settings/settings.page';
+
+export const SETTINGS_ROUTES: Routes = [
+    {
+        path: '',
+        component: SettingsPage,
+    },
+];
