+++ src/app/shared/components/image-card/image-card.component.ts
@@ -1,3 +1,4 @@
+import { AsyncPipe, NgStyle } from '@angular/common';
 import {
     ChangeDetectionStrategy,
     Component,
@@ -7,11 +8,20 @@ import {
 } from '@angular/core';
 
 import { AbstractImageCardComponent } from '@app/shared/components/abstract-image-card/abstract-image-card.component';
+import { IsImageWidthLargerPipe } from '@app/shared/pipes/is-image-width-larger/is-image-width-larger.pipe';
+import { SkeletonBlockComponent } from '@app/shared/components/skeleton-block/skeleton-block.component';
 
 @Component({
     selector: 'app-image-card',
     templateUrl: './image-card.component.html',
     styleUrls: ['./image-card.component.scss'],
+    standalone: true,
+    imports: [
+        NgStyle,
+        AsyncPipe,
+        IsImageWidthLargerPipe,
+        SkeletonBlockComponent,
+    ],
     changeDetection: ChangeDetectionStrategy.OnPush,
     encapsulation: ViewEncapsulation.None,
 })
