+++ src/app/shared/components/skeleton-block/skeleton-block.component.ts
@@ -4,11 +4,16 @@ import {
     Input,
     ViewEncapsulation,
 } from '@angular/core';
+import { NgStyle } from '@angular/common';
 
 @Component({
     selector: 'app-skeleton-block',
     templateUrl: './skeleton-block.component.html',
     styleUrls: ['./skeleton-block.component.scss'],
+    standalone: true,
+    imports: [
+        NgStyle,
+    ],
     changeDetection: ChangeDetectionStrategy.OnPush,
     encapsulation: ViewEncapsulation.None,
 })
