--- src/app/shared/components/skeleton-block/skeleton-block.module.ts
@@ -1,17 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-
-import { SkeletonBlockComponent } from '@app/shared/components/skeleton-block/skeleton-block.component';
-
-@NgModule({
-    imports: [
-        CommonModule,
-    ],
-    declarations: [
-        SkeletonBlockComponent,
-    ],
-    exports: [
-        SkeletonBlockComponent,
-    ],
-})
-export class SkeletonBlockModule {}
