+++ src/app/store/settings/reducers/settings.reducer.ts
@@ -1,7 +1,7 @@
 import { createReducer, on } from '@ngrx/store';
 
 import { SettingsStoreInterface } from '@app/store/settings/types/settings-store.interface';
-import { defaultAvailableLangs } from '@app/core/transloco-root.module';
+import { defaultAvailableLangs } from '@app/core/providers/transloco/transloco.provider';
 import {
     resetSettingsAction,
     togglePlayerModeAction,
